Extract stop point validation code within export function to a separate validator service.
